instance method is not private
Getting a SystemStackError: stack level too deep using this + rails 4.1.9

see
https://github.com/rails/rails/blob/4-1-stable/activesupport/lib/active_support/per_thread_registry.rb#L39

Not sure how to add a regression test, I tried, however I think it is
hard to test this, as active_support will always be included before
monorails as rspec will include it. And if AS is before, the problem
will not happen.